# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Edge Dispatch Framework API",
    "version": "0.5.0",
    "description": "Edge Dispatch Framework — 开源边缘分发/加速框架，通过中心调度将用户请求分配到最优边缘节点，支持 302 重定向、DNS/GSLB、网关反代等多种接入方式。"
  },
  "servers": [
    { "url": "http://localhost:8080", "description": "本地 Docker 演示环境" }
  ],
  "components": {
    "securitySchemes": {
      "EdgeAgentBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Edge Agent API key 签名验证"
      },
      "AdminJWT": {
        "type": "http",
        "scheme": "bearer",
        "description": "管理控制台 JWT (通过 /login 获取)"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": { "$ref": "#/components/schemas/ErrorDetail" }
        }
      },
      "ErrorDetail": {
        "type": "object",
        "properties": {
          "code":    { "type": "string", "description": "错误码" },
          "message": { "type": "string", "description": "错误消息" },
          "request_id": { "type": "string", "description": "请求追踪 ID" },
          "details": { "type": "object", "additionalProperties": { "type": "string" } }
        }
      },
      "NodeStatus": {
        "type": "string",
        "enum": ["REGISTERED", "ACTIVE", "DEGRADED", "QUARANTINED", "OFFLINE", "DISABLED", "MAINTENANCE"]
      },
...
      "Endpoint": {
        "type": "object",
        "properties": {
          "scheme": { "type": "string" },
          "host":   { "type": "string" },
          "port":   { "type": "integer" }
        }
      },
      "Capabilities": {
        "type": "object",
        "properties": {
          "inbound_reachable": { "type": "boolean" },
          "cache_disk_gb":     { "type": "integer", "format": "int64" },
          "max_uplink_mbps":   { "type": "integer", "format": "int64" },
          "supports_https":    { "type": "boolean" },
          "nat_type":          { "type": "string" },
          "tunnel_required":   { "type": "boolean" }
        }
      },
      "NodeScores": {
        "type": "object",
        "properties": {
          "reachable_score": { "type": "number", "format": "double" },
          "health_score":    { "type": "number", "format": "double" },
          "risk_score":      { "type": "number", "format": "double" }
        }
      },
      "NodeRuntime": {
        "type": "object",
        "properties": {
          "cpu":           { "type": "number", "format": "double" },
          "mem_mb":        { "type": "integer", "format": "int64" },
          "disk_free_gb":  { "type": "integer", "format": "int64" },
          "conn":          { "type": "integer", "format": "int64" }
        }
      },
      "NodeTraffic": {
        "type": "object",
        "properties": {
          "egress_mbps":  { "type": "number", "format": "double" },
          "ingress_mbps": { "type": "number", "format": "double" },
          "err_5xx_rate": { "type": "number", "format": "double" }
        }
      },
      "NodeCache": {
        "type": "object",
        "properties": {
          "hit_ratio": { "type": "number", "format": "double" }
        }
      },
      "ContentSummary": {
        "type": "object",
        "properties": {
          "node_id":      { "type": "string" },
          "hot_keys":     { "type": "array", "items": { "type": "string" } },
          "bloom_filter": { "type": "string", "format": "byte" },
          "bloom_k":      { "type": "integer" },
          "total_keys":   { "type": "integer", "format": "int64" },
          "updated_at":   { "type": "integer", "format": "int64" }
        }
      },
      "Node": {
        "type": "object",
        "properties": {
          "node_id":        { "type": "string" },
          "tenant_id":      { "type": "string" },
          "project_id":     { "type": "string" },
          "name":           { "type": "string" },
          "region":         { "type": "string" },
          "isp":            { "type": "string" },
          "asn":            { "type": "string" },
          "status":         { "$ref": "#/components/schemas/NodeStatus" },
          "weight":         { "type": "integer" },
          "last_seen_at":   { "type": "string", "format": "date-time" },
          "created_at":     { "type": "string", "format": "date-time" },
          "updated_at":     { "type": "string", "format": "date-time" },
          "endpoints":      { "type": "array", "items": { "$ref": "#/components/schemas/Endpoint" } },
          "capabilities":   { "$ref": "#/components/schemas/Capabilities" },
          "scores":         { "$ref": "#/components/schemas/NodeScores" },
          "labels":         { "type": "object", "additionalProperties": { "type": "string" } },
          "disable_reason": { "type": "string" },
          "maintain_until": { "type": "string", "format": "date-time", "nullable": true }
        }
      },
      "NodeAuth": {
        "type": "object",
        "properties": {
          "type":  { "type": "string" },
          "token": { "type": "string" },
          "exp":   { "type": "integer", "format": "int64" }
        }
      },
      "RegisterRequest": {
        "type": "object",
        "required": ["node_name"],
        "properties": {
          "node_name":        { "type": "string", "description": "节点名称" },
          "public_endpoints": { "type": "array", "items": { "$ref": "#/components/schemas/Endpoint" } },
          "region":           { "type": "string" },
          "isp":              { "type": "string" },
          "tenant_id":        { "type": "string" },
          "project_id":       { "type": "string" },
          "capabilities":     { "$ref": "#/components/schemas/Capabilities" }
        }
      },
      "RegisterResponse": {
        "type": "object",
        "properties": {
          "node_id": { "type": "string" },
          "auth":    { "$ref": "#/components/schemas/NodeAuth" }
        }
      },
      "HeartbeatRequest": {
        "type": "object",
        "required": ["node_id"],
        "properties": {
          "node_id":         { "type": "string", "description": "节点 ID" },
          "ts":              { "type": "integer", "format": "int64" },
          "runtime":         { "$ref": "#/components/schemas/NodeRuntime" },
          "traffic":         { "$ref": "#/components/schemas/NodeTraffic" },
          "cache":           { "$ref": "#/components/schemas/NodeCache" },
          "content_summary": { "$ref": "#/components/schemas/ContentSummary" }
        }
      },
      "ClientInfo": {
        "type": "object",
        "properties": {
          "ip":     { "type": "string", "description": "客户端 IP (为空时自动填充)" },
          "region": { "type": "string" },
          "isp":    { "type": "string" }
        }
      },
      "ResourceInfo": {
        "type": "object",
        "required": ["key"],
        "properties": {
          "type":   { "type": "string" },
          "key":    { "type": "string", "description": "资源标识" },
          "scheme": { "type": "string" }
        }
      },
      "ClientHints": {
        "type": "object",
        "properties": {
          "need_range": { "type": "boolean" }
        }
      },
      "DispatchRequest": {
        "type": "object",
        "required": ["resource"],
        "properties": {
          "client":   { "$ref": "#/components/schemas/ClientInfo" },
          "resource": { "$ref": "#/components/schemas/ResourceInfo" },
          "hints":    { "$ref": "#/components/schemas/ClientHints" }
        }
      },
      "DispatchToken": {
        "type": "object",
        "properties": {
          "type":  { "type": "string" },
          "value": { "type": "string" },
          "exp":   { "type": "integer", "format": "int64" }
        }
      },
      "CandidateMeta": {
        "type": "object",
        "properties": {
          "region":      { "type": "string" },
          "isp":         { "type": "string" },
          "proxy_mode":  { "type": "string", "enum": ["direct", "tunnel"] },
          "gateway_url": { "type": "string" }
        }
      },
      "Candidate": {
        "type": "object",
        "properties": {
          "id":       { "type": "string" },
          "endpoint": { "type": "string" },
          "weight":   { "type": "integer" },
          "meta":     { "$ref": "#/components/schemas/CandidateMeta" }
        }
      },
      "DispatchResponse": {
        "type": "object",
        "properties": {
          "request_id": { "type": "string" },
          "ttl_ms":     { "type": "integer", "format": "int64" },
          "token":      { "$ref": "#/components/schemas/DispatchToken" },
          "candidates": { "type": "array", "items": { "$ref": "#/components/schemas/Candidate" } }
        }
      },
...
      "NodeAdminPatch": {
        "type": "object",
        "properties": {
          "labels":         { "type": "object", "additionalProperties": { "type": "string" } },
          "weight":         { "type": "integer" },
          "project_id":     { "type": "string" },
          "tenant_id":      { "type": "string" },
          "disabled":       { "type": "boolean" },
          "disable_reason": { "type": "string" },
          "disable_until":  { "type": "string", "format": "date-time" }
        }
      },
...
      "DisableNodeRequest": {
        "type": "object",
        "required": ["reason"],
        "properties": {
          "reason":  { "type": "string", "description": "关闭原因" },
          "message": { "type": "string" },
          "until":   { "type": "string", "format": "date-time", "description": "维护截止时间" }
        }
      }
    }
  },
  "paths": {
...
    "/v1/nodes/register": {
      "post": {
        "tags": ["Edge Agent"],
        "summary": "注册边缘节点",
        "security": [{ "EdgeAgentBearer": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "注册成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/RegisterResponse" }
              }
            }
          },
          "400": { "description": "请求参数错误", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "401": { "description": "未认证", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "500": { "description": "注册失败", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/v1/nodes/heartbeat": {
      "post": {
        "tags": ["Edge Agent"],
        "summary": "节点心跳上报",
        "security": [{ "EdgeAgentBearer": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/HeartbeatRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "example": "ok" }
                  }
                }
              }
            }
          },
          "400": { "description": "请求参数错误", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "404": { "description": "节点不存在", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "500": { "description": "心跳失败", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/v1/nodes/{nodeID}": {
      "get": {
        "tags": ["Edge Agent"],
        "summary": "查询节点信息",
        "security": [{ "EdgeAgentBearer": [] }],
        "parameters": [
          { "name": "nodeID", "in": "path", "required": true, "schema": { "type": "string" }, "description": "节点 ID" }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Node" } } }
          },
          "404": { "description": "节点不存在", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      },
      "delete": {
        "tags": ["Edge Agent"],
        "summary": "吊销节点",
        "security": [{ "EdgeAgentBearer": [] }],
        "parameters": [
          { "name": "nodeID", "in": "path", "required": true, "schema": { "type": "string" }, "description": "节点 ID" }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "status": { "type": "string", "example": "revoked" } }
                }
              }
            }
          },
          "404": { "description": "节点不存在", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "500": { "description": "吊销失败", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/v1/dispatch/resolve": {
      "post": {
        "tags": ["Edge Agent"],
        "summary": "调度分发（返回最优边缘节点）",
        "security": [{ "EdgeAgentBearer": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/DispatchRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "分发结果",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DispatchResponse" }
              }
            }
          },
          "400": { "description": "请求参数错误", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "403": { "description": "IP 被封禁", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "429": { "description": "请求被限流", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "500": { "description": "分发失败", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
...
    "/internal/admin/v1/nodes/{nodeID}": {
      "get": {
        "tags": ["Admin - 节点"],
        "summary": "获取节点详情",
        "security": [{ "AdminJWT": [] }],
        "parameters": [
          { "name": "nodeID", "in": "path", "required": true, "schema": { "type": "string" }, "description": "节点 ID" }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Node" } } }
          },
          "404": { "description": "节点不存在", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      },
      "patch": {
        "tags": ["Admin - 节点"],
        "summary": "更新节点属性",
        "security": [{ "AdminJWT": [] }],
        "parameters": [
          { "name": "nodeID", "in": "path", "required": true, "schema": { "type": "string" }, "description": "节点 ID" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/NodeAdminPatch" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status":  { "type": "string", "example": "patched" },
                    "node_id": { "type": "string" }
                  }
                }
              }
            }
          },
          "400": { "description": "请求参数错误", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "500": { "description": "更新失败", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/internal/admin/v1/nodes/{nodeID}:disable": {
      "post": {
        "tags": ["Admin - 节点"],
        "summary": "禁用节点",
        "security": [{ "AdminJWT": [] }],
        "parameters": [
          { "name": "nodeID", "in": "path", "required": true, "schema": { "type": "string" }, "description": "节点 ID" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/DisableNodeRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status":  { "type": "string", "example": "disabled" },
                    "node_id": { "type": "string" }
                  }
                }
              }
            }
          },
          "500": { "description": "操作失败", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/internal/admin/v1/nodes/{nodeID}:enable": {
      "post": {
        "tags": ["Admin - 节点"],
        "summary": "启用节点",
        "security": [{ "AdminJWT": [] }],
        "parameters": [
          { "name": "nodeID", "in": "path", "required": true, "schema": { "type": "string" }, "description": "节点 ID" }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status":  { "type": "string", "example": "enabled" },
                    "node_id": { "type": "string" }
                  }
                }
              }
            }
          },
          "500": { "description": "操作失败", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/internal/admin/v1/nodes/{nodeID}:revoke": {
      "post": {
        "tags": ["Admin - 节点"],
        "summary": "吊销节点",
        "security": [{ "AdminJWT": [] }],
        "parameters": [
          { "name": "nodeID", "in": "path", "required": true, "schema": { "type": "string" }, "description": "节点 ID" }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status":  { "type": "string", "example": "revoked" },
                    "node_id": { "type": "string" }
                  }
                }
              }
            }
          },
          "500": { "description": "操作失败", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
...
  },
...
}